% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lori_MI.R
\name{boot.lori}
\alias{boot.lori}
\title{Bootstrap function}
\usage{
boot.lori(Y, prob = 0.1)
}
\arguments{
\item{Y}{count data matrix}

\item{prob}{proportion of entries to remove in each bootstrap sample}
}
\value{
A new data set with additional missing values and Poisson noise
}
\description{
Bootstrap function
}
\examples{
Y <- matrix(rpois(25, 1:25), 5)
yboot <- boot.lori(Y)
}
