% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp2_distinct.R
\name{comp2_distinct}
\alias{comp2_distinct}
\title{Compare 2-row grouped subsets of distinct dataframe}
\usage{
comp2_distinct(data = NULL, g_by = NULL, distinct = F)
}
\description{
Compare 2-row grouped subsets of distinct dataframe
}
