% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_functions.R
\name{default_projection}
\alias{default_projection}
\title{Default projection on covariate space: center by row and column.}
\usage{
default_projection(M)
}
\arguments{
\item{X}{A matrix.}
}
\value{
The projection of X
}
\description{
Default projection on covariate space: center by row and column.
}
\examples{
X = matrix(rnorm(rep(0, 15)), 5)
default_projection(X)
}
