% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_functions.R
\name{deviance}
\alias{deviance}
\title{Computes deviance of Poisson parameter matrix with count data matrix.}
\usage{
deviance(Y, X)
}
\arguments{
\item{Y}{A matrix of counts (contingency table, same size as X).}

\item{X}{A Poisson parameter matrix.}
}
\value{
the deviance between \code{X} and \code{Y}
}
\description{
Computes deviance of Poisson parameter matrix with count data matrix.
}
\examples{
X = matrix(rnorm(rep(0, 15)), 5)
Y = matrix(rpois(length(c(X)), exp(c(X))), 5)
deviance(Y, X)
}
