% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f2dfs.R
\name{f2dfs}
\alias{f2dfs}
\title{Apply function to multiple dataframes}
\usage{
f2dfs(FUN, dfs = NULL)
}
\arguments{
\item{dfs}{character vector of names of dataframes, tibbles ideally}

\item{f}{function}
}
\description{
Use this function to apply any function to multiple dataframes specified by names
}
\examples{
##Easy example
data(iris)
data(mtcars)
vars <- c("iris","mtcars")
f2dfs(nrow,vars)
}
