% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_missing.R
\name{ggplot_missing}
\alias{ggplot_missing}
\title{Function to plot missing map}
\usage{
ggplot_missing(x, split = F, by = 25)
}
\arguments{
\item{x}{should be a dataframe. As of now it is not being checked so use with care.}

\item{split}{should the dataframe be split into smaller chunks}

\item{by}{if split is TRUE we are splitting the dataframe into chunks with by number of columns}
}
\description{
It calls an internal function basic_miss_plot that shouldn't be used outside of this.
If split is TRUE then multiple graphs are produced on chunks of dataframe.
This might improve visibility in large dataframes
}
\examples{
data("airquality")
ggplot_missing(airquality)

## splitting into 3 graphs
ggplot_missing(airquality,T,2)
}
