% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_regularization.R
\name{lambda_QUT_covariates}
\alias{lambda_QUT_covariates}
\title{Computes the threshold $\lambda_{\text{QUT}}$.}
\usage{
lambda_QUT_covariates(Y, projection = default_projection, quantile = 0.95,
  n = 10000)
}
\arguments{
\item{Y}{A matrix of counts (contingency table).}
}
\value{
the null estimator under constraint Theta = 0
}
\description{
Computes the threshold $\lambda_{\text{QUT}}$.
}
\examples{
X = matrix(rnorm(rep(0, 15)), 5)
Y = matrix(rpois(length(c(X)), exp(c(X))), 5)
res = estimate_null_model(Y)
}
