% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lori_plots.R
\name{legend.col}
\alias{legend.col}
\title{legend.col}
\usage{
legend.col(col, lev)
}
\arguments{
\item{col}{color}

\item{lev}{vector of levels}
}
\description{
legend.col
}
\examples{
image(matrix(rnorm(18), nrow=9), col = grey(seq(1, 0, length = 100)))
legend.col(grey(seq(0, 1, length = 256)), lev = seq(0,100,length.out = 3))
}
