% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lori.R
\docType{package}
\name{Lori-package}
\alias{Lori}
\alias{Lori-package}
\title{Lori: Personal library of useful functions}
\description{
Personal library of useful functions. Use with care.
}
\keyword{internal}
