% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lori_pal.R
\name{lori_pal}
\alias{lori_pal}
\title{Default Lori palettes}
\usage{
lori_pal(pal = 1)
}
\arguments{
\item{pal}{choose wanted palette. So far only 5 ale implemented.}
}
\description{
Default Lori palettes
}
\examples{
lori_pal(1)
plot_colours(lori_pal(1))

library(gridExtra)
g <- purrr::map(1:6,~plot_colours(lori_pal(.x)))
do.call("gridExtra::grid.arrange", c(g, nrow=1))
}
