% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persp_jk.R
\name{persp_jk}
\alias{persp_jk}
\title{Plot 3d surfaces, similar to matlab surf function}
\usage{
persp_jk(
  x = NULL,
  y = NULL,
  z,
  colormap = viridisLite::viridis(256),
  theta = -30,
  phi = 30,
  expand = 0.5,
  ticktype = "detailed",
  legend = T,
  ...
)
}
\arguments{
\item{x, y}{locations of grid lines at which the values in z are measured. These must be in ascending order. By default, equally spaced values from 0 to 1 are used. If x is a list, its components x$x and x$y are used for x and y, respectively.}

\item{z}{a matrix containing the values to be plotted (NAs are allowed).}

\item{colormap}{colormap used for the z-values}

\item{theta, phi}{angles defining the viewing direction. theta gives the azimuthal direction and phi the colatitude.}

\item{expand}{a expansion factor applied to the z coordinates. Often used with 0 < expand < 1 to shrink the plotting box in the z direction.}

\item{ticktype}{haracter: "simple" draws just an arrow parallel to the axis to indicate direction of increase; "detailed" draws normal ticks as per 2D plots.}

\item{legend}{should the legend be drawn. The default is TRUE.}
}
\description{
Plot 3d surfaces, similar to matlab surf function
}
\examples{
data(ari)
persp_jk(z = ari)

## changing parameters
x <- seq(1,50,1)
y <- seq(1,20,1)
data(parula64)
persp_jk(x,y,ari,colormap = parula64,xlab = "t", ylab= "K neighbors",zlab = "Adjusted Rand Index",legend=F)

## changing border color
persp_jk(x,y,ari,colormap = viridisLite::viridis(64,option = "A"),xlab = "t", ylab= "K neighbors",zlab = "Adjusted Rand Index", border = "white")

## generating terrain
require(ambient)
require(scico)
set.seed(4425346)
terrain <- noise_simplex(c(100,100))
persp_jk(z = terrain, colormap = scico(256, palette = "vik"), border = "NA", legend = F)
}
