% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_palette_jk.R
\name{pick_palette_jk}
\alias{pick_palette_jk}
\title{Pick color palette from a picture}
\usage{
pick_palette_jk(
  path = NULL,
  k = 3,
  plot_image = F,
  algorithm = "agnes",
  grid = 60,
  full_object = T,
  linkage = "single",
  kmax = 10
)
}
\arguments{
\item{path}{Path used to get the .jpg picture. Before running this function you can run file.choose() function to get the desired path.}

\item{k}{Desired number of colors in a palette.}

\item{plot_image}{Should the original image be plotted. The default value is FALSE.}

\item{algorithm}{Two methods are implemented. The currently available are "agnes", "diana" and "kmeans". The default is "agnes".}

\item{grid}{The size of grid used for pixel sampling.}

\item{linkage}{Link used in hierarchical clustering. Default is "single".}

\item{kmax}{Maximal k wanted in the tree while using diana.}
}
\description{
This function uses clustering on a subset of pixels from given picture. The final output is mean of pixels in each cluster converted to hex.
Agnes is computationaly much more intensive than kmeans. Hierarchical clustering is primarily used to construct the whole dendrogram.
So we can just ouput the whole tree and let you plot or get colors afterwards. Use function \code{\link{tree2color}} from this package.
However if you choose the kmeans algorithm, you can only get one vector of colors.
}
\examples{
path <- "./man/figures/duck.jpg"
tree <- pick_palette_jk(path) #keep in mind this can take up to 5 mins
colours <- tree2color(tree, k = 10)
plot_colours(colours)

## using diana
tree <- pick_palette_jk(path,algorithm = "diana")
tree2color(tree, k = 10) \%>\% sort_colours \%>\% plot_colours

## or use something like this
color_list <- purrr::map(1:10,~tree2color(tree,.x))
# do.call(pals::pal.bands,color_list) # commented out in cas pals fails to install
#compare this with
plot_colours(rgb(pick_palette_jk(path,algorithm = "kmeans",k = 10)))
#hierarchical clustering has great potential but is really slow compared to kmeans


}
