% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lori_plots.R
\name{plot_cov}
\alias{plot_cov}
\title{plot_cov}
\usage{
plot_cov(res_lori)
}
\arguments{
\item{res_lori}{lori output}
}
\description{
plot_cov
}
\examples{
Y <- matrix(rpois(9, lambda=1:9),nrow=3)
Y[sample(1:9, 2)] <- NA
cov <- matrix(rnorm(18), nrow=9)
colnames(cov) <- c("cov1","cov2")
res <- lori(Y,cov,1,1,rank.max=2,maxit=1,size=2)
plot_cov(res)
}
