% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpart2r_vec.R
\name{rpart2r_vec}
\alias{rpart2r_vec}
\title{Metaprogramming function to return vectorized R code from rpart object}
\usage{
rpart2r_vec(frame, f_depth = 1, f_row_order = 1, sep_char = "\\t")
}
\arguments{
\item{frame}{This should be rpart$frame that has been supplied by prepare_rpart4code.R}

\item{f_depth}{Please don't change this. It's only meant to change itself during recursion.}

\item{f_row_order}{Please don't change this. It's only meant to change itself during recursion.}
}
\description{
Metaprogramming function to return vectorized R code from rpart object
}
\examples{
prepare_rpart4code(tree_2010_2015) \%>\% rpart2r() \%>\% message()
}
