% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_colours.R
\name{sort_colours}
\alias{sort_colours}
\title{Function to sort colors}
\usage{
sort_colours(col = NULL, method = "PCA_HSV")
}
\arguments{
\item{col}{color vector in hex format}

\item{method}{only "PCA_HSV" is implemented}
}
\description{
Warning it is not clear how to sort colors properly.
As of now, only one way is implemented here.
First we transform the hex values to RGB and then to HSV.
Then we apply pca retaining only 1 componen and sort it.
}
