% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm_fillmap.R
\name{svm_fillmap}
\alias{svm_fillmap}
\title{Replace colors in SVM fill plot}
\usage{
svm_fillmap(g, fillmap = NULL, s_lim = NULL)
}
\arguments{
\item{g}{g should be an object output from `plot_svm_jk()`` where fill_plot = T. It's crucial that it's from that function because I embedded some information into the g$s_lim_jk.}

\item{fillmap}{This should be a diverging color palette. Default is using just `scale_fill_gradient2()`}

\item{s_lim}{This is the upper limit for the new scale. The lower limit is -s_lim. You can set this to a specific number, but it defaults to the highest number needed for the plot.}
}
\description{
Replace colors in SVM fill plot
}
\examples{
data(df)
library(e1071)
model <- svm(label ~ x1 + x2, data = df, kernel = "radial", gamma = 1, scale = F, cost = 1.8)
g <- plot_svm_jk(df,model,fill_plot = T, plot_contour = F, plot_decision = F)
svm_fillmap(g)
}
