% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree2color.R
\name{tree2color}
\alias{tree2color}
\title{Cut a "colored" tree}
\usage{
tree2color(tree_model, k = 3, fun = "mean")
}
\arguments{
\item{tree_model}{model produced by \code{\link{pick_palette_jk}} function.}

\item{k}{number of clusters.}

\item{fun}{function used on grouped data. Perhaps it makes sense to use different functions than mean. Median might be a good choice. Input should be character.}
}
\description{
This function is used for getting hard clusters by cutting a tree. Use function \code{\link{plot_colours}} on the result of this function to get a plot.
}
\examples{
path <- "./man/figures/duck.jpg"
tree <- pick_palette_jk(path)
colours <- tree2color(tree, k = 10)
plot_colours(colours)

 ## or use something like this
 color_list <- purrr::map(1:10,~tree2color(tree,.x))
 do.call(pals::pal.bands,color_list)
}
