% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_functions.R
\name{default_projection}
\alias{default_projection}
\title{Default projection on covariate space: center by row and column.}
\usage{
default_projection(M)
}
\arguments{
\item{M}{A matrix.}
}
\value{
M centered by row and column
}
\description{
Default projection on covariate space: center by row and column.
}
\examples{
\dontshow{
X = matrix(rnorm(rep(0, 15)), 5)
proj_X <- default_projection(X)
}
}
