% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggraph_jk.R
\name{ggraph_jk}
\alias{ggraph_jk}
\title{Plot undirected graphs constructed from 2D data}
\usage{
ggraph_jk(
  data = NULL,
  matrix_g = NULL,
  x = colnames(data)[1],
  y = colnames(data)[2],
  label = colnames(data)[3],
  colormap = NULL,
  color = "edge",
  plot_points = T,
  color_points = NULL,
  color_lines = NULL,
  cex_p = 3,
  size = 0.5,
  matlab = F
)
}
\arguments{
\item{data}{The data to be displayed in this layer.}

\item{matrix_g}{Matrix of edges in calculated graph. Could be a full matrix
or sparse from the Matrix package.}

\item{x}{Name of the first coordinate.}

\item{y}{Name of the second coordinate.}

\item{label}{Name of the column coresponding to ground truth in clustering.
Could be omitted.}

\item{colormap}{Colormap to color the vertices/edges of a graph.}

\item{color}{A string indicating of what to color, using a colormap. Four
options available: "none", "edge", "degree" and "wdegree". Degree uses
colormap to color points based on the number of edges going to/from the
given vertex. Wdegree calculates the sum of weights of edges in each
vertex.}

\item{plot_points}{boolean. Should we plot the vertices?}

\item{color_points}{The chosen color for coloring points.}

\item{color_lines}{The chosen color for coloring edges.}

\item{cex_p}{Change the size of points.}

\item{size}{Change the width of edges.}

\item{matlab}{boolean. Should we use default MATLAB colors and try to
replicate the ggplot theme?}
}
\description{
Plot undirected graphs constructed from 2D data
}
\examples{
## Edge coloring
data(spiraly)
data(Wfull)
ggraph_jk(spiraly,Wfull)

ggraph_jk(spiraly,Wfull,colormap = colormap::colormap(colormap = "hot", n = 256),cex_p = 1.5)

## exporting graph and changing it
g <- ggraph_jk(data = spiraly, matrix_g = Wfull, color = "degree",cex_p = 3)
g + theme(legend.position = c(0.9,0.2),legend.background=element_blank())

##coloring nodes based on the sum of all weights of edges connected to them.
g <- ggraph_jk(data = spiraly, matrix_g = Wfull, color = "wdegree",cex_p = 3,matlab = T,color_lines = "black")
g + theme(legend.position = c(0.9,0.2),legend.background=element_blank())
}
