% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lori-package.R
\docType{package}
\name{lori-package}
\alias{lori-package}
\title{lori: Imputation of High-Dimensional Count Data using Side Information}
\description{
Analysis, imputation, and multiple imputation of count data using covariates. LORI uses a log-linear Poisson model where main row and column effects, as well as effects of known covariates and interaction terms can be fitted. The estimation procedure is based on the convex optimization of the Poisson loss penalized by a Lasso type penalty and a nuclear norm. LORI returns estimates of main effects, covariate effects and interactions, as well as an imputed count table. The package also contains a multiple imputation procedure. The methods are described in Robin, Josse, Moulines and Sardy (2019) \href{https://arxiv.org/abs/1703.02296v4}{arXiv:1703.02296v4}.
}
\author{
\strong{Maintainer}: Genevieve Robin \email{genevieve.robin@cnrs.fr}

}
\keyword{internal}
