% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lori_pal.R
\name{lori_pal}
\alias{lori_pal}
\title{Default Lori palettes}
\usage{
lori_pal(pal = 1)
}
\arguments{
\item{pal}{choose wanted palette. So far only 3 ale implemented. 1 implies the default matlab palette. 2 and 3 are specific to this package.}
}
\description{
Default Lori palettes
}
